Refactor to factory method pattern
