+++ apphost.settings.json
@@ -9,8 +9,9 @@
 
   "LlmProvider": "MicrosoftFoundry",
 
+  "AgentMode": "Single",
+
   "GitHub": {
-    "Endpoint": "https://models.github.ai/inference",
     "Token": "{{GITHUB_PAT}}",
     "Model": "openai/gpt-5-mini"
   },
@@ -29,5 +30,9 @@
       "ApiKey": "{{MICROSOFT_FOUNDRY_PROJECT_API_KEY}}",
       "DeploymentName": "model-router"
     }
+  },
+
+  "GitHubCopilot": {
+    "Token": "{{GITHUB_PAT}}"
   }
 }
