+++ src/InterviewCoach.Agent/Constants.cs
@@ -2,5 +2,7 @@ namespace InterviewCoach.Agent;
 
 public class Constants
 {
+    public const string AgentMode = "AgentMode";
     public const string LlmProvider = "LlmProvider";
+    public const string GitHubToken = "GITHUB_TOKEN";
 }
