+++ src/InterviewCoach.AppHost/AppHost.cs
@@ -2,11 +2,6 @@ var builder = DistributedApplication.CreateBuilder(args);
 
 // var foundry = builder.AddBicepTemplate("foundry", "../../infra/foundry.bicep");
 
-// GitHub Token — required for Mode 3 (Multi-Agent Handoff with GitHub Copilot SDK).
-// Set the value in apphost.settings.json under "Parameters:github-token",
-// as a user secret, or enter it in the Aspire Dashboard when prompted.
-var githubToken = builder.AddParameter(ResourceConstants.GitHubToken, secret: true);
-
 var mcpMarkItDown = builder.AddDockerfile(ResourceConstants.McpMarkItDown, "../InterviewCoach.Mcp.MarkItDown/packages/markitdown-mcp")
                            .WithExternalHttpEndpoints()
                            .WithImageTag("latest")
@@ -23,9 +18,8 @@ var mcpInterviewData = builder.AddProject<Projects.InterviewCoach_Mcp_InterviewD
 
 var agent = builder.AddProject<Projects.InterviewCoach_Agent>(ResourceConstants.Agent)
                    .WithExternalHttpEndpoints()
-                   .WithLlmReference(builder.Configuration)
+                   .WithLlmReference(builder.Configuration, args)
                    .WithEnvironment(ResourceConstants.LlmProvider, builder.Configuration[ResourceConstants.LlmProvider] ?? string.Empty)
-                   .WithEnvironment("GITHUB_TOKEN", githubToken)
                    .WithReference(mcpMarkItDown.GetEndpoint("http"))
                    .WithReference(mcpInterviewData)
                    .WaitFor(mcpMarkItDown)
