+++ src/InterviewCoach.AppHost/appsettings.json
@@ -9,8 +9,9 @@
 
   "LlmProvider": "MicrosoftFoundry",
 
+  "AgentMode": "Single",
+
   "GitHub": {
-    "Endpoint": "https://models.github.ai/inference",
     "Token": "{{GITHUB_PAT}}",
     "Model": "openai/gpt-5-mini"
   },
@@ -27,7 +28,11 @@
     "Project": {
       "Endpoint": "{{MICROSOFT_FOUNDRY_PROJECT_ENDPOINT}}",
       "ApiKey": "{{MICROSOFT_FOUNDRY_PROJECT_API_KEY}}",
-      "DeploymentName": "model-router"
+      "DeploymentName": "gpt-5-mini"
     }
+  },
+
+  "GitHubCopilot": {
+    "Token": "{{GITHUB_PAT}}"
   }
 }
